+++ dejafu/Test/DejaFu/Conc/Internal/Memory.hs
@@ -129,7 +129,7 @@ writeBarrier (WriteBuffer wb) = mapM_ flush $ M.elems wb where
 -- | Add phantom threads to the thread list to commit pending writes.
 addCommitThreads :: WriteBuffer r -> Threads n r -> Threads n r
 addCommitThreads (WriteBuffer wb) ts = ts <> M.fromList phantoms where
-  phantoms = [ (ThreadId Nothing $ negate tid, mkthread c)
+  phantoms = [ (ThreadId (Id Nothing $ negate tid), mkthread c)
              | ((_, b), tid) <- zip (M.toList wb) [1..]
              , c <- maybeToList (go $ viewl b)
              ]